requirements.txt
sympy==1.11.1
Flask==2.2.2
Pillow==9.2.0